+++ Plugin/Source/Processors/Timing_Effects/FlutterProcess.cpp
@@ -1,37 +1,38 @@
 #include "FlutterProcess.h"
 
-void FlutterProcess::prepare (double sampleRate, int samplesPerBlock)
+void FlutterProcess::prepare (double sampleRate, int samplesPerBlock, int numChannels)
 {
     fs = (float) sampleRate;
 
-    for (int ch = 0; ch < 2; ++ch)
+    depthSlew.resize ((size_t) numChannels);
+    for (auto& dSlew : depthSlew)
     {
-        depthSlew[ch].reset (sampleRate, 0.05);
-        depthSlew[ch].setCurrentAndTargetValue (depthSlewMin);
-
-        phase1[ch] = 0.0f;
-        phase2[ch] = 0.0f;
-        phase3[ch] = 0.0f;
+        dSlew.reset (sampleRate, 0.05);
+        dSlew.setCurrentAndTargetValue (depthSlewMin);
     }
+
+    phase1.resize ((size_t) numChannels, 0.0f);
+    phase2.resize ((size_t) numChannels, 0.0f);
+    phase3.resize ((size_t) numChannels, 0.0f);
 
     amp1 = -230.0f * 1000.0f / fs;
     amp2 = -80.0f * 1000.0f / fs;
     amp3 = -99.0f * 1000.0f / fs;
     dcOffset = 350.0f * 1000.0f / fs;
 
-    flutterBuffer.setSize (2, samplesPerBlock);
+    flutterBuffer.setSize (numChannels, samplesPerBlock);
 }
 
-void FlutterProcess::prepareBlock (float curDepth, float flutterFreq, int numSamples)
+void FlutterProcess::prepareBlock (float curDepth, float flutterFreq, int numSamples, int numChannels)
 {
-    depthSlew[0].setTargetValue (jmax (depthSlewMin, curDepth));
-    depthSlew[1].setTargetValue (jmax (depthSlewMin, curDepth));
+    for (auto& dSlew : depthSlew)
+        dSlew.setTargetValue (jmax (depthSlewMin, curDepth));
 
-    angleDelta1 = MathConstants<float>::twoPi * 1.0f * flutterFreq / fs;
-    angleDelta2 = MathConstants<float>::twoPi * 2.0f * flutterFreq / fs;
-    angleDelta3 = MathConstants<float>::twoPi * 3.0f * flutterFreq / fs;
+    angleDelta1 = MathConstants<float>::twoPi * flutterFreq / fs;
+    angleDelta2 = 2.0f * angleDelta1;
+    angleDelta3 = 3.0f * angleDelta1;
 
-    flutterBuffer.setSize (2, numSamples, false, false, true);
+    flutterBuffer.setSize (numChannels, numSamples, false, false, true);
     flutterBuffer.clear();
     flutterPtrs = flutterBuffer.getArrayOfWritePointers();
 }